Add files via upload
A basic Xylophone app created using Flutter is a simple mobile application that allows users to play musical notes by tapping on different colored bars displayed on the screen. The app's user interface typically consists of a column of different colored bars, each representing a musical note. When a user taps on a bar, the app produces the corresponding musical note.

The Xylophone app is built using Flutter, a popular open-source framework for developing mobile applications. Flutter uses the Dart programming language and provides a fast and efficient way to build cross-platform apps for Android and iOS.

The Xylophone app makes use of the audio_player package to play sounds, which allows the app to produce the musical notes when a user taps on a bar. The app also makes use of Flutter's Material Design framework to provide a clean and intuitive user interface.